Reapply r196551 which was accidentally reverted by r223637 (update to OpenBSD pf 4.5).
Fix argument ordering to memcpy as well as the size of the copy in the
(theoretical) case that pfi_buffer_cnt should be greater than ~_max.

This fix the failure when you hit the self table size and force it to be
resized.

MFC after:	3 days
Sponsored by:	Rubicon Communications (Netgate)


git-svn-id: https://svn.freebsd.org/base/head@287119 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f