Be more GCC-friendly with attributes
Being clang the default compiler, we were always giving precedence to
the __has_attribute check. Unfortunately clang generally doesn't support
the new attributes (alloc_size was briefly supported and then reverted)
so we were always doing both checks. Give the precedence to GCC as that is
the working case now.

Do the same for  __has_builtin() for consistency.


git-svn-id: https://svn.freebsd.org/base/head@287254 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f