Relax serseq option operation for reads.
Previously, with serseq enabled, next command was unblocked only after
previous completed.  With this change, for read operations, next command
is unblocked as soon as last media read completed.  This is important
for frontends that actually wait for data move completion (like camtgt),
or when data are moved through the HA link, or especially when both.


git-svn-id: https://svn.freebsd.org/base/head@287967 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f