META_MODE: Avoid command changing in 2nd build.
If the command to be ran changes then a rebuild is caused. Checking
exists(${DESTDIR}...) from make results in this on the 2nd and
possibly subsequent builds due to staging during build.  Avoid this
by always running the existence check in the make sh command.

Sponsored by:	EMC / Isilon Storage Division


git-svn-id: https://svn.freebsd.org/base/head@287972 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f