The Sun RPC framework uses a netbuf structure to represent the transport specific form of a universal transport address.  The structure is expected to be opaque to consumers.  In the current implementation, the structure contains a pointer to a buffer that holds the actual address.
In rpcbind(8), netbuf structures are copied directly, which would
result in two netbuf structures that reference to one shared
address buffer.  When one of the two netbuf structures is freed,
access to the other netbuf structure would result in an undefined
result that may crash the rpcbind(8) daemon.

Fix this by making a copy of the buffer that is going to be freed
instead of doing a shallow copy.

Security:	FreeBSD-SA-15:24.rpcbind
Security:	CVE-2015-7236


git-svn-id: https://svn.freebsd.org/base/head@288383 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f